+++ scripts/install-postgres.sh
@@ -2,6 +2,15 @@
 
 # PostgreSQL
 
+if [ -f /home/vagrant/.hercules-installed/postgresql ]
+then
+    echo "PostgreSQL is already installed"
+    exit 0
+fi
+
+touch /home/vagrant/.hercules-installed/postgresql
+
+
 # prepare
 sudo echo 'deb http://apt.postgresql.org/pub/repos/apt/ xenial-pgdg main' >> /etc/apt/sources.list.d/pgdg.list
 sudo wget --quiet -O - https://www.postgresql.org/media/keys/ACCC4CF8.asc | sudo apt-key add -
