+++ scripts/install-rabbit.sh
@@ -2,6 +2,15 @@
 
 # RabbitMQ with Erlang
 
+if [ -f /home/vagrant/.hercules-installed/rabbitmq ]
+then
+    echo "RabbitMQ is already installed"
+    exit 0
+fi
+
+touch /home/vagrant/.hercules-installed/rabbitmq
+
+
 # install Erlang
 sudo echo "deb http://binaries.erlang-solutions.com/debian xenial contrib" > /etc/apt/sources.list.d/eslerlang.list
 wget -O - http://binaries.erlang-solutions.com/debian/erlang_solutions.asc | sudo apt-key add -
