+++ scripts/install-rethinkdb.sh
@@ -2,6 +2,15 @@
 
 # RethinkDB
 
+if [ -f /home/vagrant/.hercules-installed/rethinkdb ]
+then
+    echo "RethinkDB is already installed"
+    exit 0
+fi
+
+touch /home/vagrant/.hercules-installed/rethinkdb
+
+
 # install
 source /etc/lsb-release && echo "deb http://download.rethinkdb.com/apt $DISTRIB_CODENAME main" | sudo tee /etc/apt/sources.list.d/rethinkdb.list
 wget -qO- https://download.rethinkdb.com/apt/pubkey.gpg | sudo apt-key add -
